Resolve circular dependencies
